:zap: Include clear session confirmation
