+++ components/validateWallet.tsx
@@ -14,41 +14,36 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 
 //MUI Components
 import Modal from "@mui/material/Modal";
 import Typography from "@mui/material/Typography";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import ContentCopyIcon from "@mui/icons-material/ContentCopy";
-import Tooltip from "@mui/material/Tooltip";
 
-//lib
-import { clearWallet, getMnemonic } from "@/lib/storage";
+//lib and custom components
+import { getMnemonic } from "@/lib/storage";
 import Image from "next/image";
-import { CustomPasswordInput, PassPhrase } from "./CustomPasswordInput";
-import { shortenAddress } from "@/lib/helper";
+import { CustomPasswordInput } from "./CustomPasswordInput";
+import { AppContext } from "@/context/appContext";
+import { ClearPassphrase } from "./ClearPassphrase";
 
-export const ValidateWallet = ({
-  open,
-  handleClose,
-  passphrase,
-  setPassphrase,
-  walletAddr,
-  setWalletAddr,
-}: {
-  open: boolean;
-  handleClose: (arg?: string) => void;
-  passphrase: PassPhrase;
-  setPassphrase: (arg: any) => void;
-  walletAddr: string | null;
-  setWalletAddr: any;
-}) => {
-  const [forgotPassphrase, setForgotPassphrase] = useState<boolean>(false);
-  const [tooltiptext, setTooltiptext] = useState("Click to Copy");
+export const ValidateWallet = () => {
   const [errorMessage, setErrorMessage] = useState("");
+  const context = useContext(AppContext);
+  if (context === undefined) {
+    throw new Error("Must be inside of a App Provider");
+  }
+  const {
+    openValidateModal,
+    passphrase,
+    setPassphrase,
+    handleCloseValidate,
+    forgotPassphrase,
+    setForgotPassphrase,
+    clearPassphrase,
+  }: any = context;
 
   const validate = (e: any) => {
     e.preventDefault();
@@ -57,32 +52,11 @@ export const ValidateWallet = ({
       if (!mnemonic || mnemonic.message === "Malformed UTF-8 data") {
         setErrorMessage("Invalid passphrase, please retry!");
       } else if (typeof mnemonic === "string") {
-        handleClose(mnemonic);
+        handleCloseValidate(mnemonic);
       }
     } else {
       setErrorMessage("Enter a passphrase!");
     }
-  };
-
-  const clearPassprase = () => {
-    setForgotPassphrase(false);
-    setPassphrase((prev: PassPhrase) => ({ ...prev, password: "" }));
-  };
-
-  const clearSession = () => {
-    clearWallet();
-    clearPassprase();
-    handleClose();
-    setWalletAddr(null);
-  };
-
-  const copyAddress = (address: string) => {
-    document.querySelector(".copyToClipboard");
-    navigator.clipboard.writeText(address);
-    setTooltiptext(`Copied: ${address}`);
-    setTimeout(() => {
-      setTooltiptext("Click to Copy");
-    }, 500);
   };
 
   useEffect(() => {
@@ -93,10 +67,10 @@ export const ValidateWallet = ({
 
   return (
     <Modal
-      open={open}
+      open={openValidateModal}
       onClose={() => {
-        clearPassprase();
-        handleClose();
+        clearPassphrase();
+        handleCloseValidate();
       }}
       aria-labelledby="modal-modal-title"
       aria-describedby="modal-modal-description"
@@ -116,259 +90,129 @@ export const ValidateWallet = ({
           maxHeight: "100%",
         }}
       >
-        {forgotPassphrase && (
-          <Typography
+        {forgotPassphrase ? (
+          <ClearPassphrase />
+        ) : (
+          <Box
             sx={{
-              fontSize: "18px",
-              fontWeight: 700,
-              display: "flex",
-              alignItems: "center",
-              cursor: "pointer",
-              width: "fit-content",
-              marginBottom: "25px",
-            }}
-            onClick={() => {
-              setForgotPassphrase(false);
+              paddingInline: "1rem",
+              width: "85%",
+              marginX: "auto",
+              "@media (max-width: 800px)": {
+                width: "90%",
+                paddingInline: "4rem",
+              },
+              "@media (max-width: 501px)": {
+                width: "100%",
+                padding: 0,
+              },
             }}
           >
-            <ChevronLeftIcon sx={{ fontSize: "30px" }} />
-            CANCEL
-          </Typography>
-        )}
-        <Box
-          sx={{
-            paddingInline: "1rem",
-            width: "85%",
-            marginX: "auto",
-            "@media (max-width: 800px)": {
-              width: "90%",
-              paddingInline: "4rem",
-            },
-            "@media (max-width: 501px)": {
-              width: "100%",
-              padding: 0,
-            },
-          }}
-        >
-          {forgotPassphrase ? (
-            <>
+            <Box
+              sx={{
+                marginBottom: "30px",
+                display: "flex",
+                columnGap: "10px",
+                "@media(min-width:800px)": {
+                  paddingLeft: "4rem",
+                },
+              }}
+            >
               <Box
                 sx={{
-                  marginBottom: "30px",
-                  display: "flex",
-                  columnGap: "10px",
-                  "@media(min-width:800px)": {
-                    paddingLeft: "4rem",
+                  marginTop: "20px",
+                }}
+              >
+                <Image
+                  src="/password.png"
+                  alt="password"
+                  width="35"
+                  height="35"
+                />
+              </Box>
+              <Box>
+                <Typography
+                  sx={{
+                    fontSize: "22px",
+                    fontWeight: 700,
+                  }}
+                >
+                  Enter Passphrase
+                </Typography>
+
+                <Typography
+                  sx={{
+                    fontSize: "16px",
+                    fontWeight: 500,
+                    marginBottom: "10px",
+                    color: "primary.dark",
+                  }}
+                >
+                  Enter the passphrase to unlock this wallet - Wallet needs to
+                  be unlocked every time page is refreshed.
+                </Typography>
+                <Typography
+                  sx={{
+                    fontSize: "12px",
+                    fontStyle: "italic",
+                  }}
+                >
+                  The passphrase is only stored locally. If you have forgetten
+                  it, you will need to clear and re-enter the mnemonic.
+                </Typography>
+              </Box>
+            </Box>
+            <form>
+              <Box
+                sx={{
+                  width: "50%",
+                  marginX: "auto",
+                  "@media (max-width: 501px)": {
+                    width: "80%",
                   },
                 }}
               >
-                <Box
-                  sx={{
-                    marginTop: "20px",
-                  }}
-                >
-                  <Image
-                    src="/password.png"
-                    alt="password"
-                    width="55"
-                    height="55"
-                  />
-                </Box>
-                <Box>
-                  <Typography
-                    sx={{
-                      fontSize: "22px",
-                      fontWeight: 700,
-                    }}
-                  >
-                    Clear Stored Wallet
-                  </Typography>
-                  <Box
-                    sx={{
-                      display: "flex",
-                      alignItems: "center",
-                      background: "#EEEEEE",
-                      border: "1px solid",
-                      borderColor: "secondary.dark",
-                      borderRadius: "3px",
-                      width: "fit-content",
-                      padding: "0.3rem 1rem",
-                      marginY: "10px",
-                    }}
-                  >
-                    <Typography
-                      fontSize={"1rem"}
-                      textAlign={"center"}
-                      fontWeight={500}
-                    >
-                      {shortenAddress(walletAddr || "", 10)}
-                    </Typography>
-                    <Tooltip
-                      title={tooltiptext}
-                      placement="top"
-                      arrow
-                      sx={{
-                        cursor: "pointer",
-                        marginLeft: "0.5rem",
-                      }}
-                    >
-                      <ContentCopyIcon
-                        sx={{
-                          marginLeft: "0.4rem",
-                          fontSize: "0.9rem",
-                          opacity: 0.7,
-                          cursor: "pointer",
-                          transition: "all .3s ease",
-                          ["&:hover"]: {
-                            opacity: 1,
-                          },
-                          ["@media(max-width:600px)"]: {
-                            fontSize: "1.5rem",
-                          },
-                        }}
-                        onClick={() => copyAddress(walletAddr || "")}
-                      />
-                    </Tooltip>
-                  </Box>
-                  <Typography
-                    sx={{
-                      fontSize: "16px",
-                      fontWeight: 500,
-                      marginBottom: "10px",
-                      color: "primary.dark",
-                    }}
-                  >
-                    Clearing the mnemonic disconnects the wallet entirely and a
-                    mnemonic will need to be re-entered to connect back to a
-                    wallet.
-                  </Typography>
-                </Box>
-              </Box>
-
-              <Box sx={{ textAlign: "center", marginBlock: "40px" }}>
-                <Button
-                  variant="contained"
-                  sx={{
-                    backgroundColor: "primary.dark",
-                    color: "#FFFFFF",
-                    "&:hover": {
-                      backgroundColor: "primary.dark",
-                    },
-                  }}
-                  onClick={clearSession}
-                >
-                  CLEAR STORED WALLET
-                </Button>
-              </Box>
-            </>
-          ) : (
-            <>
-              <Box
-                sx={{
-                  marginBottom: "30px",
-                  display: "flex",
-                  columnGap: "10px",
-                  "@media(min-width:800px)": {
-                    paddingLeft: "4rem",
-                  },
-                }}
-              >
-                <Box
-                  sx={{
-                    marginTop: "20px",
-                  }}
-                >
-                  <Image
-                    src="/password.png"
-                    alt="password"
-                    width="35"
-                    height="35"
-                  />
-                </Box>
-                <Box>
-                  <Typography
-                    sx={{
-                      fontSize: "22px",
-                      fontWeight: 700,
-                    }}
-                  >
-                    Enter Passphrase
-                  </Typography>
-
-                  <Typography
-                    sx={{
-                      fontSize: "16px",
-                      fontWeight: 500,
-                      marginBottom: "10px",
-                      color: "primary.dark",
-                    }}
-                  >
-                    Enter the passphrase to unlock this wallet - Wallet needs to
-                    be unlocked every time page is refreshed.
-                  </Typography>
+                <CustomPasswordInput
+                  passphrase={passphrase}
+                  setPassphrase={setPassphrase}
+                />
+                {errorMessage && (
                   <Typography
                     sx={{
                       fontSize: "12px",
-                      fontStyle: "italic",
+                      color: "error.main",
                     }}
                   >
-                    The passphrase is only stored locally. If you have forgetten
-                    it, you will need to clear and re-enter the mnemonic.
+                    {errorMessage}
                   </Typography>
-                </Box>
+                )}
               </Box>
-              <form>
-                <Box
-                  sx={{
-                    width: "50%",
-                    marginX: "auto",
-                    "@media (max-width: 501px)": {
-                      width: "80%",
-                    },
-                  }}
+              <Box sx={{ textAlign: "center", marginBlock: "40px" }}>
+                <Button
+                  type="submit"
+                  variant="contained"
+                  disabled={!passphrase.password}
+                  onClick={validate}
                 >
-                  <CustomPasswordInput
-                    passphrase={passphrase}
-                    setPassphrase={setPassphrase}
-                  />
-                  {errorMessage && (
-                    <Typography
-                      sx={{
-                        fontSize: "12px",
-                        color: "error.main",
-                      }}
-                    >
-                      {errorMessage}
-                    </Typography>
-                  )}
-                </Box>
-                <Box sx={{ textAlign: "center", marginBlock: "40px" }}>
-                  <Button
-                    type="submit"
-                    variant="contained"
-                    disabled={!passphrase.password}
-                    onClick={validate}
-                  >
-                    Enter
-                  </Button>
-                </Box>
-              </form>
-              <Typography
-                sx={{
-                  fontSize: "12px",
-                  fontWeight: 600,
-                  fontStyle: "italic",
-                  cursor: "pointer",
-                  textAlign: "center",
-                  textDecoration: "underline",
-                }}
-                onClick={() => setForgotPassphrase(true)}
-              >
-                Forgot Passphrase? Clear Mnemonic
-              </Typography>
-            </>
-          )}
-        </Box>
+                  Enter
+                </Button>
+              </Box>
+            </form>
+            <Typography
+              sx={{
+                fontSize: "12px",
+                fontWeight: 600,
+                fontStyle: "italic",
+                cursor: "pointer",
+                textAlign: "center",
+                textDecoration: "underline",
+              }}
+              onClick={() => setForgotPassphrase(true)}
+            >
+              Forgot Passphrase? Clear Mnemonic
+            </Typography>
+          </Box>
+        )}
       </Box>
     </Modal>
   );
